+++ src/services/api/user/checkUser.js
@@ -1,10 +1,10 @@
-import api from '../../lib/api';
+import api from '../../../lib/api';
 
-const userCheckAPI = (email) => {
+const checkUserAPI = (email) => {
   return api.get(
     `${process.env.REACT_APP_SERVER_DOMAIN}/api/user/check?email=${email}`,
     { withCredentials: true },
   );
 };
 
-export default userCheckAPI;
+export default checkUserAPI;
